🐛 FIX: PR feedback implemented, separate cloud run v2 block infra to models directory
