--- docs/cloud_run_v2.md
@@ -1,1 +1,0 @@
-::: prefect_gcp.cloud_run_v2
